+++ src/utils/index.ts
@@ -2,3 +2,4 @@ export * from './csv-parser';
 export * from './category-service';
 export * from './text-utils';
 export * from './date-utils';
+export * from './text-normalization';